refactor: move all components from index to specific files
