+++ src/components/app/userActionFormPetitionSignature/constants.ts
@@ -1,0 +1,1 @@
+export const DEFAULT_INNER_WIDTH_CLASS_NAME = 'max-md: mx-auto md:w-[75%] md:min-w-[540px]'
